chore: create gtagHelper with tests
